requirements.txt
pymongo[srv]==4.2.0
pandas
scikit-learn
PyYAML
scipy
dill
imblearn
flask
-e .